example/requirements.txt
django
djangorestframework
djangorestframework-deepserializer